price_events/requirements.txt
nameko-kafka
pymongo
pytest